requirements.txt
requests>=2.31.0
sounddevice>=0.4.6
numpy>=1.24.0
pystray>=0.19.5
Pillow>=10.0.0
customtkinter>=5.2.0
keyboard>=0.13.5
pyperclip>=1.8.2
pygame>=2.5.0
pyinstaller>=6.0.0
